overlay improvements
