+++ src/main/java/graphics/devMode/DevWindow.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.ArrayList;
 import java.awt.event.ActionListener;
 
+import graphics.common.overlays.MenuOverlay;
+import graphics.common.overlays.OverlayButtonListener;
 import org.jetbrains.annotations.NotNull;
 
 import graphicsEngine.windows.AbstractPage;
@@ -27,7 +29,7 @@ public class DevWindow extends CommonWindow {
         //setActivePage(StartingPage.getStaticPageKey());
         //setActivePage(PlayPage.getStaticPageKey());
 
-        //setOverlay(new MenuOverlay(new OverlayButtonListener(this)));
+        setOverlay(new MenuOverlay(new OverlayButtonListener(this)));
     }
 
     /**
